backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
gpxpy==1.6.2
python-dotenv==1.0.1
supabase==2.7.4
